feat: show upcoming votes when active vote
