+++ components/Votes/index.ts
@@ -1,1 +1,4 @@
 export { Votes } from "./Votes";
+export { ActiveVotes } from "./ActiveVotes";
+export { PastVotes } from "./PastVotes";
+export { UpcomingVotes } from "./UpcomingVotes";
